# Document/11_openapi_starter.yaml
openapi: 3.0.3
info:
  title: DoodhDirect API
  version: 1.0.0
  description: >-
    Farm-to-home dairy platform API. Application-owned business timestamps use
    India-local wall-clock values in Asia/Kolkata. Dairy business timestamps are
    accepted from clients as UTC instants and converted to India-local wall-clock
    values at the API boundary; responses serialize them without a Z suffix or
    offset. Provider/external and deferred infrastructure timestamps retain UTC
    semantics and their existing UTC-suffixed field names.
servers:
  - url: https://api.example.com/api/v1
paths:
...
  /auth/send-otp:
    post:
      summary: Send an identity OTP
      description: Anonymous. Accepts mobile and OtpPurpose (0 login, 1 registration, 2 password reset, 3 employee invitation). The provider request id (reqId) returned by the response must be supplied to retry-otp and verify-otp.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendOtpRequest' }
      responses:
        '200':
          description: OTP request accepted envelope carrying the MSG91 reqId
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SendOtpResultEnvelope' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '429': { description: OTP request rate limited }
  /auth/retry-otp:
    post:
      summary: Resend an identity OTP
      description: Anonymous. Re-issues the OTP for an existing challenge identified by mobile and purpose. A new reqId is returned and replaces the previous one on the challenge.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendOtpRequest' }
      responses:
        '200':
          description: OTP request accepted envelope carrying the new reqId
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SendOtpResultEnvelope' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { description: OTP request rate limited }
  /auth/verify-otp:
    post:
      summary: Verify an identity OTP and complete the auth action
      description: Anonymous. The backend validates the entered code with the MSG91 widget provider using the reqId returned by send-otp or retry-otp, then validates the provider access token server-side before issuing the authenticated session.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyOtpRequest' }
      responses:
        '200':
          description: Authenticated session envelope
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSessionEnvelope' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { description: OTP verification rate limited }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication is required or the access token is invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: The owned resource was not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        field: { type: string, nullable: true }
        message: { type: string }
    ErrorResponse:
      type: object
      required: [success, errors]
      properties:
        success: { type: boolean, enum: [false] }
        data: { nullable: true }
        message: { type: string, nullable: true }
        errors:
          type: array
          items: { $ref: '#/components/schemas/ApiError' }
    OtpPurpose:
      type: integer
      format: int32
      enum: [0, 1, 2, 3]
      description: 0 login, 1 registration, 2 password reset, 3 employee invitation.
    SendOtpRequest:
      type: object
      required: [mobile, purpose]
      properties:
        mobile:
          type: string
          minLength: 10
          maxLength: 15
          description: Normalized India mobile number.
        purpose:
          $ref: '#/components/schemas/OtpPurpose'
    SendOtpResultEnvelope:
      type: object
      required: [success, data, errors]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          type: object
          required: [reqId]
          properties:
            reqId:
              type: string
              maxLength: 64
              description: MSG91 provider request id. The client holds it only in memory and must supply it to retry-otp and verify-otp.
        message: { type: string, nullable: true }
        errors:
          type: array
          maxItems: 0
          items: { $ref: '#/components/schemas/ApiError' }
    VerifyOtpRequest:
      type: object
      required: [mobile, code, purpose, reqId, device]
      properties:
        mobile:
          type: string
          minLength: 10
          maxLength: 15
        code:
          type: string
          minLength: 6
          maxLength: 6
          description: Six-digit code entered by the user.
        purpose:
          $ref: '#/components/schemas/OtpPurpose'
        reqId:
          type: string
          maxLength: 64
          description: The reqId returned by send-otp or retry-otp for this challenge.
        device:
          $ref: '#/components/schemas/OtpDeviceInfo'
    OtpDeviceInfo:
      type: object
      required: [deviceIdentifier]
      properties:
        deviceIdentifier: { type: string, minLength: 1, maxLength: 200 }
        deviceName: { type: string, nullable: true }
        platform: { type: string, nullable: true }
    AuthSessionEnvelope:
      type: object
      required: [success, data, errors]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          type: object
          description: Full authenticated session (user, roles, permissions, branch ids, token pair) returned after successful OTP verification. The backend validates the MSG91 access token server-side before issuing this session.
        message: { type: string, nullable: true }
        errors:
          type: array
          maxItems: 0
          items: { $ref: '#/components/schemas/ApiError' }
...